+++ Lineairealgebra/Vector.h
@@ -14,19 +14,19 @@ public:
 	void add(Vector& vector);
 	void subtract(Vector& vector);
 
-	int x() const;
-	void setX(int x);
-	int y() const;
-	void setY(int y);
-	int pixelX() const;
-	int pixelY() const;
+	double x() const;
+	void setX(double x);
+	double y() const;
+	void setY(double y);
+	double pixelX() const;
+	double pixelY() const;
 
 	Vector* next() const;
 	void setNext(Vector& next);
 
 private:
-	int x_;
-	int y_;
+	double x_;
+	double y_;
 
 	Vector* next_;
 };
